// src/app/enums/caminhos.enum.ts
export enum Caminhos {
  chifre = 'Chifre',
  artesao = 'Artesão',
  paje = 'Pajé',

  guardiao = 'Guardião Gélido',
  emissario = 'Emissário dos Antigos',
  sacerdote = 'Sacerdote de Nagari',

  presa = 'Presa Furiosa',
  arauto = 'Arauto da Forja',
  mao = 'Mão de Caapora',

  sentinela = 'Sentinela',
  sussurro = 'Sussurro',
  sombrio = 'Sombrio',

  pescador = 'Pescador Indomável',
  construtor = 'Iça-velas',
  navegador = 'Navega-Ondas',

  Bravo = 'Bravo',
  Cronista = 'Cronista',
  sabio = 'Sábio',

}